requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
certifi==2024.2.2
dj-database-url==2.1.0
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.25.2
idna==3.7
mistralai==0.1.8
orjson==3.10.1
packaging==24.0
psycopg2==2.9.9
pydantic==2.7.0
pydantic_core==2.18.1
PyJWT==2.8.0
python-dotenv==1.0.1
replicate==0.25.2
sniffio==1.3.1
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
whitenoise==6.6.0
